refactored file structure again back to the original. I am big idiot did not know how to make modules.
